PythonHub/pyproject.toml
[project]
name = "ue5-zeromq-hub"
version = "1.0.0"
description = "ZeroMQ communication hub for Unreal Engine 5 and MATLAB/Simulink integration"
authors = [
    {name = "UE5PyCom Team"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "pyzmq==27.0.1",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "pillow==11.3.0",
    "requests==2.32.4",
    "PySide6==6.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pyinstaller>=6.15.0",
    "pyinstaller-hooks-contrib>=2025.8"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
